added navbar to the request rider page
